Configured offlineimap to fetch my mail into the Maildir in ~/.Mail.
Its rc isn’t included, because there is too much personal data in there.
But there is a sample in .offlineimap/ which should make it easy to add
accounts.
Keys are saved via python-gnomekeyring and the helper script.
Most scripts are snagged from the internet.
Yet there is one improvement for i3: .offlineimap/run_offlineimap.sh
runs oi and counts new and unread messages in all accounts afterwards,
which is then saved to .offlineimap/.new_mail.
An (unfinished) wrapper for i3status fetches the numbers and displays
them in the status bar.
TODO: Colors!