Added command line arguments to the offlineimap script and fixed a bug.
run_offlineimap.sh now accepts -q and -b#.
-q:  Run a quick sync.
-b#: Run a backup with Back in Time (backup job number #, should ideally
include the maildir ~/.offlineimap.
Fixed a bug that removed the .new_mail file while executing the script.
Now actually removing the .new_mail file when there is no new mail,
fixed the i3 mailwrapper .py to not crash when the .new_mail file is N/A
but rather pipe an empty string to i3bar.